Refactor live_streams router into smaller, more focused routers
